requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-CORS==4.0
PyMySQL==1.1
python-dotenv==1.0.1
Werkzeug==3.0.3
requests==2.32.3